Fix '&' multiplication
Let'S do it by example:

  .ident_a, .ident_b {
    & + & + & { color: red; }
  }

We have to generate all permutations of the parent identifier list and
the amount of ampersand child identifiers. In the above case, the
exploded identifier list has ident_count**ampersand_count entries (8).

As a bonus, we get the same identifier sort order as lessc does.

Conflicts:
	lesscpy/plib/identifier.py