Bump version to 0.10
Moved to lesscpy.__version__ and use it in setup.py. Also, don't put the
whole license text into the 'license' tag but rather the short form
'MIT'. The license file is part of the source distribution already and
was added to package_data (so that it ends up somewhere when 'setup.py
install' is invoked).